Refactored features into a separate sub-package
